Grey line added on 1 ans 2 tabs. Formatting and spinners fixed on 2 tab. Firat tab summ - long!!

+++ app/src/main/java/com/by/alex/depositcalcupd/Formatter.java
@@ -14,7 +14,7 @@ public class Formatter {
         return myFormat.format(f).replace(","," ");
     }
 
-    public String formatSumm (int f) {
+    public String formatSumm (long f) {
         DecimalFormatSymbols s = new DecimalFormatSymbols();
         //s.setDecimalSeparator('.');
         return summFormat.format(f).replace(",", " ");
@@ -32,8 +32,8 @@ public class Formatter {
         return number;
     }
 
-    public int parseSumm(String str) {
-        int number = Integer.parseInt(str.replace(",", "").replace(" ",""));
+    public long parseSumm(String str) {
+        long number = Long.parseLong(str.replace(",", "").replace(" ",""));
 
         return number;
     }
